Move settings function
