+++ apps/vs-code-designer/src/app/utils/appSettings/localSettings.ts
@@ -11,6 +11,7 @@ import { writeFormattedJson } from '../fs';
 import { parseJson } from '../parseJson';
 import { DialogResponses, parseError } from '@microsoft/vscode-azext-utils';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
+import { MismatchBehavior } from '@microsoft/vscode-extension';
 import type { ILocalSettingsJson } from '@microsoft/vscode-extension';
 import * as fse from 'fs-extra';
 import * as path from 'path';
@@ -120,4 +121,42 @@ export async function getLocalSettingsJson(
       AzureWebJobsStorage: '',
     },
   };
+}
+
+/**
+ * Set local.settings.json values.
+ * @param {IActionContext} context - Command context.
+ * @param {string} logicAppPath - Project path.
+ * @param {string} key - Key to be updated.
+ * @param {string} value - Value to be updated.
+ * @param {MismatchBehavior} behavior - Behaviour of the update.
+ */
+export async function setLocalAppSetting(
+  context: IActionContext,
+  logicAppPath: string,
+  key: string,
+  value: string,
+  behavior: MismatchBehavior = MismatchBehavior.Prompt
+): Promise<void> {
+  const localSettingsPath: string = path.join(logicAppPath, localSettingsFileName);
+  const settings: ILocalSettingsJson = await getLocalSettingsJson(context, localSettingsPath);
+
+  settings.Values = settings.Values || {};
+  if (settings.Values[key] === value) {
+    return;
+  } else if (settings.Values[key]) {
+    if (behavior === MismatchBehavior.Prompt) {
+      const message: string = localize('SettingAlreadyExists', "Local app setting '{0}' already exists. Overwrite?", key);
+      if (
+        (await context.ui.showWarningMessage(message, { modal: true }, DialogResponses.yes, DialogResponses.cancel)) !== DialogResponses.yes
+      ) {
+        return;
+      }
+    } else if (behavior === MismatchBehavior.DontChange) {
+      return;
+    }
+  }
+
+  settings.Values[key] = value;
+  await writeFormattedJson(localSettingsPath, settings);
 }
